backend/requirements.txt
flask
flask-cors
flask-socketio
opencv-python-headless
ultralytics
gunicorn
cvzone
python-dotenv
# Production WSGI worker (gthread mode for MJPEG streaming)
eventlet==0.37.1
